requirements.txt
gradio
moviepy==1.0.3
Pillow
opencv-python
numpy
